Create shared Core Navigator, which encapsulates both HistoryRecorderHandle and Navigator.
