+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Fixed a potential hang when `PassiveLocationDataSource` fails to download routing tiles. ([#2808](https://github.com/mapbox/mapbox-navigation-ios/pull/2808))
 * Fixed an issue where `RouteController` or `PassiveLocationDataSource` sometimes snapped the user’s location assuming a path that violated a turn restriction. ([#2808](https://github.com/mapbox/mapbox-navigation-ios/pull/2808))
 * Improved performance and decreased memory usage when downloading routing tiles. ([#2808](https://github.com/mapbox/mapbox-navigation-ios/pull/2808))
+* `PassiveLocationManager.startUpdatingLocation(completionHandler:)` no longer requires optional `completionHandler:` property. ([#2823](https://github.com/mapbox/mapbox-navigation-ios/pull/2823))
 
 ### CarPlay
