+++ Sources/MapboxCoreNavigation/PassiveLocationDataSource.swift
@@ -16,7 +16,7 @@ open class PassiveLocationDataSource: NSObject {
      - parameter directions: The directions service that allows the location data source to access road network data. If this argument is omitted, the shared `Directions` object is used.
      - parameter systemLocationManager: The location manager that provides raw locations for the receiver to match against the road network.
      
-     - postcondition: Call `startUpdatingLocation(completionHandler:)` afterwards to begin receiving location updates.
+     - postcondition: Call `startUpdatingLocation()` afterwards to begin receiving location updates.
      */
     public required init(directions: Directions = Directions.shared, systemLocationManager: NavigationLocationManager? = nil) {
         self.directions = directions
@@ -41,8 +41,8 @@ open class PassiveLocationDataSource: NSObject {
     /**
      The underlying navigator that performs map matching.
      */
-    var navigator: Navigator {
-        return Navigator.shared
+    var navigator: MapboxNavigationNative.Navigator {
+        return Navigator.shared.navigator
     }
     
     /**
